% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coNBMTX_main.R
\name{coNBMTX}
\alias{coNBMTX}
\title{Conditional negative binomial regression for sample-paired mgx and mtx data}
\usage{
coNBMTX(datamat, group_formula, normal_method = "CSS")
}
\arguments{
\item{datamat}{A list contains mtx, mgx and metadata.}

\item{group_formula}{Specify the group information.}

\item{normal_method}{Specify the way to estimate the scaling factor.}
}
\value{
A list contains the analysis result of each gene.
}
\description{
Conditional negative binomial regression for sample-paired mgx and mtx data
}
\examples{
mtxtable <- matrix(rnbinom(5000,size=10,mu=50),nrow=100)
mgxtable <- matrix(rnbinom(5000,size=5,mu=30),nrow=100)
metatable <- data.frame(ID=paste("S",seq(1:50),sep=""),group=c(rep("a",25),rep("b",25)))
datamat = list(mtx=mtxtable,mgx=mgxtable,metadata=metatable)
group_formula = "~group"
output = coNBMTX(datamat=datamat,group_formula=group_formula)

}
