% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coNBMTX_main.R
\name{GetTab}
\alias{GetTab}
\title{Get a table for the analysis result of coNBMTX}
\usage{
GetTab(output)
}
\arguments{
\item{output}{The output of coNBMTX.}
}
\value{
A dataframe contains the analysis result of all genes.
}
\description{
Get a table for the analysis result of coNBMTX
}
\examples{
mtxtable <- matrix(rnbinom(5000,size=10,mu=50),nrow=100)
mgxtable <- matrix(rnbinom(5000,size=5,mu=30),nrow=100)
metatable <- data.frame(ID=paste("S",seq(1:50),sep=""),group=c(rep("a",25),rep("b",25)))
datamat = list(mtx=mtxtable,mgx=mgxtable,metadata=metatable)
group_formula = "~group"
output = coNBMTX(datamat=datamat,group_formula=group_formula)
Tab = GetTab(output)

}
