% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coNBMTX_main.R
\name{TopTab}
\alias{TopTab}
\title{Get the top table for the analysis result of coNBMTX}
\usage{
TopTab(output, groupname)
}
\arguments{
\item{output}{The output of coNBMTX.}

\item{groupname}{The name of the group you want to check.}
}
\value{
A dataframe contains the analysis result of top 10 genes.
}
\description{
Get the top table for the analysis result of coNBMTX
}
